# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LitXplore",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/analysis/{paper_id}/analyze": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Analyze Paper",
        "description": "Analyze a paper and generate At-a-Glance summary.\n\n- Generates or refreshes the At-a-Glance analysis immediately.\n- Returns PaperAnalysis (in_depth may be null).\n\nArgs:\n    paper_id: The ID of the paper to analyze\n    force_refresh: If True, bypass cache and regenerate analysis\n    current_user: Authenticated user\n\nReturns:\n    PaperAnalysis with at_a_glance populated",
        "operationId": "analyzePaper",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "paper_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Paper Id"
            }
          },
          {
            "name": "force_refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperAnalysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/{paper_id}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Get Paper Analysis",
        "description": "Retrieve the latest PaperAnalysis (including cached In-Depth if computed).\n\nArgs:\n    paper_id: The ID of the paper\n    current_user: Authenticated user\n\nReturns:\n    PaperAnalysis with all available data",
        "operationId": "getPaperAnalysis",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "paper_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Paper Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperAnalysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/{paper_id}/in-depth": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Compute In Depth",
        "description": "Trigger In-Depth Analysis computation for comprehensive understanding of paper sections.\n\n- Generates detailed explanations for all major paper sections\n- Returns PaperAnalysis with in_depth populated\n\nArgs:\n    paper_id: The ID of the paper\n    current_user: Authenticated user\n\nReturns:\n    PaperAnalysis with in_depth populated",
        "operationId": "computeInDepth",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "paper_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Paper Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperAnalysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/papers/{paper_id}/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Chat With Paper",
        "description": "Chat endpoint for paper-specific questions.\n\nStreams responses using Server-Sent Events (SSE) format.\n\nArgs:\n    paper_id: The ID of the paper to chat about\n    message: The user's question or message\n    current_user: Authenticated user\n\nReturns:\n    StreamingResponse with SSE-formatted chunks",
        "operationId": "chatWithPaper",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "paper_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Paper Id"
            }
          },
          {
            "name": "message",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The user's question or message",
              "title": "Message"
            },
            "description": "The user's question or message"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AtAGlanceAnalysis": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "authors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Authors"
          },
          "affiliations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affiliations"
          },
          "abstract": {
            "type": "string",
            "title": "Abstract"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords"
          },
          "introduction": {
            "type": "string",
            "title": "Introduction"
          },
          "related_work": {
            "type": "string",
            "title": "Related Work"
          },
          "problem_statement": {
            "type": "string",
            "title": "Problem Statement"
          },
          "methodology": {
            "type": "string",
            "title": "Methodology"
          },
          "results": {
            "type": "string",
            "title": "Results"
          },
          "discussion": {
            "type": "string",
            "title": "Discussion"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          },
          "future_work": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Future Work"
          },
          "conclusion": {
            "type": "string",
            "title": "Conclusion"
          }
        },
        "type": "object",
        "required": [
          "title",
          "authors",
          "affiliations",
          "abstract",
          "keywords",
          "introduction",
          "related_work",
          "problem_statement",
          "methodology",
          "results",
          "discussion",
          "limitations",
          "future_work",
          "conclusion"
        ],
        "title": "AtAGlanceAnalysis"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InDepthAnalysis": {
        "properties": {
          "introduction": {
            "type": "string",
            "title": "Introduction"
          },
          "related_work": {
            "type": "string",
            "title": "Related Work"
          },
          "problem_statement": {
            "type": "string",
            "title": "Problem Statement"
          },
          "methodology": {
            "type": "string",
            "title": "Methodology"
          },
          "results": {
            "type": "string",
            "title": "Results"
          },
          "discussion": {
            "type": "string",
            "title": "Discussion"
          },
          "limitations": {
            "type": "string",
            "title": "Limitations"
          },
          "conclusion_future_work": {
            "type": "string",
            "title": "Conclusion Future Work"
          }
        },
        "type": "object",
        "required": [
          "introduction",
          "related_work",
          "problem_statement",
          "methodology",
          "results",
          "discussion",
          "limitations",
          "conclusion_future_work"
        ],
        "title": "InDepthAnalysis",
        "description": "Comprehensive, detailed analysis of each paper section."
      },
...
      "PaperAnalysis": {
        "properties": {
          "paper": {
            "$ref": "#/components/schemas/PaperMetadata"
          },
          "at_a_glance": {
            "$ref": "#/components/schemas/AtAGlanceAnalysis"
          },
          "in_depth": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InDepthAnalysis"
              },
              {
                "type": "null"
              }
            ]
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0.0"
          },
          "model_tag": {
            "type": "string",
            "title": "Model Tag",
            "default": "gemini-2.0-flash"
          }
        },
        "type": "object",
        "required": [
          "paper",
          "at_a_glance",
          "generated_at"
        ],
        "title": "PaperAnalysis"
      },
      "PaperMetadata": {
        "properties": {
          "paper_id": {
            "type": "string",
            "title": "Paper Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "authors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Authors"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "source": {
            "type": "string",
            "enum": [
              "upload",
              "arxiv",
              "url"
            ],
            "title": "Source",
            "default": "upload"
          }
        },
        "type": "object",
        "required": [
          "paper_id",
          "title",
          "authors"
        ],
        "title": "PaperMetadata"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}